pyproject.toml
[tool.poetry]
name = "Youtube video and audio-transcription-processor"
version = "0.1.0"
description = ""
authors = ["deduu <dedy.ariansyah1@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
colorama = "0.4.6"
filelock = "3.16.1"
fsspec = "2024.10.0"
idna = "3.10"
jinja2 = "3.1.4"
llvmlite = "0.43.0"
markupsafe = "3.0.2"
more-itertools = "10.5.0"
mpmath = "1.3.0"
networkx = "3.4.2"
numba = "0.60.0"
numpy = "2.0.2"
openai-whisper = "20240930"
regex = "2024.11.6"
requests = "2.32.3"
sympy = "1.13.1"
tiktoken = "0.8.0"


torch = "2.6.0+cu124"
torchvision = "0.21.0+cu124"
torchaudio = "2.6.0+cu124"

tqdm = "4.67.1"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
yt-dlp = "^2025.2.19"
whisper = "^1.1.10"
pyannote-audio = "^3.3.2"
einops = "^0.8.1"
streamlit = "^1.42.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
